Record both source and visual loactions
Prior to this commit, only the source line and column was recorded when
parsing a box drawing.  This was not enough information to reverse the
process (i.e. represent the graph as a box drawing).

This change introduces a new type to be used as the graph node type,
which represents both the source and visual locations of the characters
in the parsed string.