Introduce location region
Prior to this change, edges of the same brush that emanated from the
same character were in indistinguishable.  For example, a bare cross,
'X', has a '/' starting at the top right and ending in the center, and
another '/' starting at center and ending at the bottom left.

The change introduces a region field to a node so that an edge can start
and end on the same character.