Ensure combining marks are ignored
Prior to this change, previous location would invariably be the
immediately preceeding code point, regardless of the visual location.
This means that the previous_location could be a combining diacritics,
which makes it problematic to ignore them.

This change updates the state's location and previous_location only when
the character is significant to the graph under construction.

Also, some of the more nuanced characters are removed from a 'clean'
string.