Ensure diagonal ports are built in the correct column
Prior to this change, a diagonal running from northwest to southeast
could jump a column if there was a similar diagonal in the subsequent
column.

This change uses an equivalence operator instead of an inequality when
comparing the current line and a port's line.  It turns out that the
inequality was only used to determine if the port should retained or
added as an edge.  Every other case required that the current line be
exactly one more than the port's line.

Also, a convenience macro has been added to provide a single code point
for every add_edge operation.