Introduce pretty_assertions
Prior to this change, the actual node list was printed prior to the
assertion that used a match.  Some these lists have become quite long
and comparing the two for a single character change is very tedious.

This change replaces the node list assertion with assert_eq from the
pretty_assertions crate for a better debugging experience.