Avoid duplicate ports
Prior to this change, the southward port of a character could be added
twice under certain conditions.  Once while processing an existing port
that is completed by the current character's northward port and then
again by the logic that handles the Initial state.

This change removes the superfluous logic in the port processing loop.